Create a directory and write out all empirical stz data

Create a directory following standard conventions to ensure appropriate placement of subdirectories and naming conventions.

Usage

dirmakR(outpath, sci_name, x, nrcs_code, regioncode, overwrite, estz_type)

Arguments

outpath

the directory folder where the final products will be placed. Note there is no need to create an independent folder to hold them, this will be included in the product.

sci_name

the scientific name of the taxon

x

Vector data product.

nrcs_code

from https://plants.usda.gov/

regioncode

a character vector string of the name for this or, the output from regioncoding

overwrite

whether to overwrite the existing directory (if it already exists)

estz_type

one of 'lg' for landscape genetics (or genetic), 'cg' for common garden, or 'cm' for climate matched. If multiple approaches choose the most robust method, e.g. 'cg' > 'lg' > 'cm'

Examples

acth7 <- sf::st_read(file.path(
  system.file(package="eSTZwritR"), "extdata", 'ACTH7.gpkg')
)
#> Reading layer `ACTH7' from data source 
#>   `/home/runner/work/_temp/Library/eSTZwritR/extdata/ACTH7.gpkg' 
#>   using driver `GPKG'
#> Simple feature collection with 4967 features and 4 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -121.5447 ymin: 36.53208 xmax: -111.0114 ymax: 46.35708
#> Geodetic CRS:  WGS 84
dirmakR('.',
  'Eriocoma thurberiana',
   acth7,
   nrcs_code = 'ACHT7',
   regioncode =  'CGB-CPN',
   estz_type = 'cg'
   )
#> Writing layer `ACHT7_cgSTZ_CGB-CPN' to data source 
#>   `./Eriocoma_thurberiana_cgSTZ/Data/Vector/ACHT7_cgSTZ_CGB-CPN.shp' using driver `ESRI Shapefile'
#> Writing 4967 features with 4 fields and geometry type Polygon.