Suggest seed zones names, or at least numeric orders, based on input polygons

This function seeks to make STZ more simple to convey, or at least visualize. Rather than using seed zone names which relate to growth measurements, phenological responses, or combinations thereof, or zones from PCA axis we bastardize the results of these analyses to two major components which humans think about extensively. Mean annual temperature, and mean annual precipitation. By rescaling these two variables together we propose intuitively simple zones that follow a gradient from hot and dry to cool and moist.

Usage

orderZones(x, SeedZone, n, ...)

Arguments

x

an sf/tibble/dataframe containing the spatial data which will be distributed as a final product

SeedZone

name of the column containing the provisional SeedZones

n

sample size of points across the entire shapefile for calculating an order, defaults to 2500. Note that this shouldn't exceed the number of raster cells of the products used to generate the stz, or results will be replicated.

...

additional parameters passed onto sf::st_sample

Examples

nc <- sf::st_read(system.file("shape/nc.shp", package="sf")) |>
  dplyr::slice_head(n = 10)
#> Reading layer `nc' from data source 
#>   `/home/runner/work/_temp/Library/sf/shape/nc.shp' using driver `ESRI Shapefile'
#> Simple feature collection with 100 features and 14 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.32385 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> Geodetic CRS:  NAD27

pts <- orderZones(nc, SeedZone = 'NAME', n = 100)
#> Warning: /home/runner/work/_temp/Library/eSTZwritR/inst/extdata/GAI-cont.tif: No such file or directory (GDAL error 4)
#> Error: [rast] file does not exist: /home/runner/work/_temp/Library/eSTZwritR/inst/extdata/GAI-cont.tif

SZplot <- ggplot2::ggplot() +
  ggplot2::geom_point(data = pts,
                     ggplot2::aes(
                       x = AnnualPrecipitation,
                       y = MeanAnnualTemp,
                       color = SeedZone)
 ) +
 ggplot2::labs(
   title = 'Relationship between Temperature and Precipitation across Seed Zones',
   color = 'Seed Zone',
   caption = 'Source: PRISM 1990-2020 climate normals') +
 ggplot2::theme_classic() +

 ggplot2::scale_y_continuous(
   name = 'Mean Annual Temperature (C\u00B0)',
   sec.axis = ggplot2::sec_axis( ~ (. * 9/5) + 32, name = "MAT (F\u00B0)")
 ) +
 ggplot2::scale_x_continuous(
   name = 'Annual Precipitation (mm)',
   sec.axis = ggplot2::sec_axis( ~ . * 0.0393701, name = "AP (inches)")
 )
#> Error: object 'pts' not found