Ensure that all fields in a shapefile are named correctly, and in the right order.

Appropriately order the fields of a shapefile and ensure consistent naming of them.

Usage

fieldsmakR(x, SeedZone, ID, SZName, AreaAcres)

Arguments

x

an sf/tibble/dataframe containing the spatial data which will be distributed as a final product

SeedZone

a string containing the numeric SeedZone code, if not supplied function will fail.

ID

a string specifying this columns names, If not supplied will create one.

SZName

a string containing a name for the SeedZone column to be used in conversation, such as 'Productivity high, Phenology late' or 'SW midmontane', if not supplied will copy 'SeedZone' column values

AreaAcres

a string containing the name for the column containing an area measurement, if not supplied this value will be calculated using epsg:5070

Examples

df <- data.frame(
id = 1:10,
gridcode = sample(1:10, replace = FALSE),
zone = sample(LETTERS, 10, replace = FALSE),
bio1_sd = runif(10, 5, 7),
bio8_mean = runif(10, 5, 7)
)

nc <- sf::st_read(system.file("shape/nc.shp", package="sf"), quiet = TRUE) |>
 dplyr::select(geometry) |>
 dplyr::slice_head(n = 10) |>
 dplyr::bind_cols(df, )

ob <- fieldsmakR(nc, SeedZone = 'gridcode')
#> There is a column(s), `zone`, which we can't figure out the purpose of.
#>       It will be returned here, but FYI a list of bioclim vars are here https://www.worldclim.org/data/bioclim.html.

dplyr::select(ob, -zone) # easily remove like so.
#> Simple feature collection with 10 features and 6 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -81.74107 ymin: 36.07282 xmax: -75.77316 ymax: 36.58965
#> Geodetic CRS:  NAD27
#>    ID SeedZone SZName AreaAcres  BIO1_SD BIO8_mean
#> 1   1        7      7  281058.5 5.062461  6.360326
#> 2   2        5      5  151003.2 5.451125  5.997691
#> 3   3        6      6  351757.7 5.601662  6.283359
#> 4   4       10     10  171630.7 6.272931  6.320569
#> 5   5        4      4  375785.3 5.958049  5.192048
#> 6   6        1      1  239133.8 5.864343  6.531200
#> 7   7        9      9  152205.6 6.412868  6.539350
#> 8   8        8      8  223298.9 6.897153  6.981425
#> 9   9        3      3  291423.0 5.360678  6.941042
#> 10 10        2      2  304632.8 5.433800  5.778366
#>                          geometry
#> 1  MULTIPOLYGON (((-81.45289 3...
#> 2  MULTIPOLYGON (((-81.17667 3...
#> 3  MULTIPOLYGON (((-80.45301 3...
#> 4  MULTIPOLYGON (((-75.94193 3...
#> 5  MULTIPOLYGON (((-77.14196 3...
#> 6  MULTIPOLYGON (((-76.7075 36...
#> 7  MULTIPOLYGON (((-76.01735 3...
#> 8  MULTIPOLYGON (((-76.46035 3...
#> 9  MULTIPOLYGON (((-78.13472 3...
#> 10 MULTIPOLYGON (((-80.02406 3...